move replay_insert & co -> mod replay
